renamed model configs simple_model.yaml -> dnatofloat_model_*pu.yaml
