+++ bin/tests/test_model/titanic_model_cpu.yaml
