+++ conf/test.config
@@ -17,8 +17,8 @@ params {
     // Input data
     csv         = "${projectDir}/bin/tests/test_data/dna_experiment/test_with_split.csv"
     exp_conf    = "${projectDir}/examples/test.json"
-    model       = "${projectDir}/bin/tests/test_model/dnatofloatmodel.py"
-    tune_conf   = "${projectDir}/bin/tests/test_model/simple_config.yaml"
+    model       = "${projectDir}/bin/tests/test_model/dnatofloat_model.py"
+    tune_conf   = "${projectDir}/bin/tests/test_model/dnatofloat_model_cpu.yaml"
 }
 
 // Limit resources so that this can run on GitHub Actions
