+++ conf/test_learn.config
@@ -15,6 +15,6 @@ params {
     csv         = "${projectDir}/bin/tests/test_data/titanic/titanic_stimulus.csv"
     exp_conf    = "${projectDir}/bin/tests/test_data/titanic/titanic_stimulus.json"  // TODO make the data transformations available to titanic data types
     model       = "${projectDir}/bin/tests/test_model/titanic_model.py"
-    tune_conf   = "${projectDir}/bin/tests/test_model/titanic_model.yaml"
+    tune_conf   = "${projectDir}/bin/tests/test_model/titanic_model_gpu.yaml"
 
 }
