+++ conf/test_stub.config
@@ -15,8 +15,8 @@ params {
     // Input data
     csv         = "${projectDir}/bin/tests/test_data/dna_experiment/test.csv"
     exp_conf    = "${projectDir}/examples/test.json"
-    model       = "${projectDir}/bin/tests/test_model/dnatofloatmodel.py"
-    tune_conf   = "${projectDir}/bin/tests/test_model/simple_config.yaml"
+    model       = "${projectDir}/bin/tests/test_model/dnatofloat_model.py"
+    tune_conf   = "${projectDir}/bin/tests/test_model/dnatofloat_model_cpu.yaml"
 }
 
 process {
